% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{cal_place_dwelling}
\alias{cal_place_dwelling}
\title{Calculate dwelling period by averaging dwelling
time between different consecutive locations}
\usage{
cal_place_dwelling(loc, time)
}
\description{
Calculate dwelling period by averaging dwelling
time between different consecutive locations
}
