% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{cart2geo.radian}
\alias{cart2geo.radian}
\title{Convert Cartesian coordinates to geopoints in radians.}
\usage{
cart2geo.radian(x)
}
\arguments{
\item{x}{A 2D vector (lat, long) representing the geo-point in radians.}
}
\description{
Convert Cartesian coordinates to geopoints in radians.
}
\seealso{
\code{\link{geo2cart.radian}}
}
