% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coordinates.R
\name{cart2geo}
\alias{cart2geo}
\title{Cartesian and geopoint conversion}
\usage{
cart2geo(x)
}
\arguments{
\item{x}{A unit-length 3D vector (x, y, z) in Cartesian system.}
}
\value{
A 2D vector (lat, long) representing the geo-point in degree.
}
\description{
Converting Cartesian coordinates into long/lat geo-points.
}
\examples{
cart2geo(c(-0.4330127, 0.7500000, 0.5000000))
}

