% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowmap.R
\name{compress_mov}
\alias{compress_mov}
\title{Compress movement history}
\usage{
compress_mov(x, y = NULL, t = NULL, gap = 6 * 3600)
}
\arguments{
\item{x,y,t}{see params of \code{\link{stcoords}}}
}
\description{
Remove duplicate location records in user's movement history.
Continuous records at the same location is merged into a single session
recording the starting and ending times.
}
\examples{
data(movement)
user_move <- subset(movement, id==1)
## 1D location
compress_mov(user_move[,c("loc", "time")])
## 2D location
compress_mov(user_move[,c("lat", "lon", "time")])
}
\seealso{
\code{\link{seq_collapsed}}
}

