% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowmap.R
\name{compress_mov2_todo}
\alias{compress_mov2_todo}
\title{TODO: there are fault when integrating with dplyr:}
\usage{
compress_mov2_todo(x, y = NULL, t = NULL, gap = 6 * 3600)
}
\description{
data(u10)
u10 %>% group_by(usr) %>% do(compress_mov_debug(x=.$site, t=.$time)) %>% filter(usr==10)
}

