% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians.}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{A number or vector of degrees.}
}
\description{
Convert degrees to radians.
}
\seealso{
\code{\link{rad2deg}}
}
