% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.iovan}
\alias{dq.iovan}
\title{Spatiotemporal data quality of Iovan's}
\usage{
dq.iovan(stcoords, type = "lonlat")
}
\arguments{
\item{type}{the selection of distance calculation function: 'lonlat' -> great circle distance
of long/lat pair, 'xy' -> euclidean distance.}

\item{x, y, t}{as like the input of \code{\link{stcoords}}. The spatial coordinates (x, y)
should be (long, lat) pair in the geographic coordinate, or quantities in Euclidean
coordinate. If (x, y) represents (long, lat) pair, the first element must be longitude.}
}
\value{
a list of data quality indicators:

theta the speed index

Q the unified quality indicator for each data point

H the entropic quality indicator for each mobility trajectory
}
\description{
In this data quality measure, Iovan et al.[1] addressed mobility data quality from both
local and global aspects. The local measure quantifies the quality of each data point,
which encodes the reasonableness of data mainly with regards to SPEED (or time and distance).
The global measure based on informational entropy quantifies the data quality within
a whole trajectory. This function implements these measures, which are also referred
to as the DYNAMIC quality for mobility data.
}
\examples{
u1 <- dplyr::filter(movement, id==1)
stc <- stcoords(u1[, c('lon','lat','time')])
dq.iovan(stc, type='lonlat')
}
\references{
[1] https://doi.org/10.1007/978-3-319-00615-4_14
}
