% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point.dynamic}
\alias{dq.point.dynamic}
\title{Spatiotemporal data quality indicator of dynamics}
\usage{
dq.point.dynamic(x, y, t)
}
\description{
Spatiotemporal data quality indicator of dynamics
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==1)
head(dq.point.dynamic(u1$lon, u1$lat, u1$time))
}

