% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point}
\alias{dq.point}
\title{Spatiotemporal data quality of stay points}
\usage{
dq.point(stcoords, pc, po, dq.min = 1e-05, na = dq.min)
}
\description{
Spatiotemporal data quality of stay points
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==1)
pc <- point.coverage(movement$lon, movement$lat)
po <- people.occurrence(movement$id, movement$lon, movement$lat)

stc <- stcoords(u1[,c('lon','lat','time')])
head(dq.point(stc, pc, po))

sessions <- gen.sessions(stc$x, stc$y, stc$t)
head(dq.point2(sessions, pc, po))
}
\seealso{
\code{\link{dq.point2}}
}

