% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point.static}
\alias{dq.point.static}
\title{Spatiotemporal data quality indicator of statics}
\usage{
dq.point.static(x, y, t, pc, po)
}
\description{
Spatiotemporal data quality indicator of statics
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==1)
pc <- point.coverage(movement$lon, movement$lat)
po <- people.occurrence(movement$id, movement$lon, movement$lat)
head(dq.point.static(u1$lon, u1$lat, u1$time, pc, po))
}

