% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.traj}
\alias{dq.traj}
\title{Spatiotemporal data quality of user trajectories}
\usage{
dq.traj(stcoords, pc, po)
}
\description{
Spatiotemporal data quality of user trajectories
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==4)
pc <- point.coverage(movement$lon, movement$lat)
po <- people.occurrence(movement$id, movement$lon, movement$lat)
stc <- stcoords(u1[,c('lon','lat','time')])
dq.traj(stc, pc, po)
dq.traj2(dq.point(stc, pc, po))
}
\seealso{
\code{\link{dq.traj2}}
}

