% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.R
\name{draw_mobility3d}
\alias{draw_mobility3d}
\title{Visualize 3D mobility data with RGL.}
\usage{
draw_mobility3d(x, y, t, group_by = NULL, col = NULL, xlab = "",
  ylab = "", tlab = "", ...)
}
\arguments{
\item{x, y}{Numeric vectors of spatial coordinates}

\item{t}{The temporal vector for each (x,y) point.}

\item{group_by}{A group indicator when multiple users are visualized.}

\item{col}{A vector of color strings. It must have the same length as unique(group_by).}

\item{xlab, ylab, tlab}{The labels for each axis.}

\item{...}{Other parameters for \code{\link[rgl]{plot3d}} or \code{\link[rgl]{axes3d}}}
}
\description{
Visualize 3D mobility data with RGL.
}
\examples{
data(movement)

users <- subset(movement, id \%in\% c(23, 20)) \%>\%
 mutate(time = time/86400 - min(time/86400)) \%>\%
 dplyr::filter(time <= 30)
 
draw_mobility3d(users$lon, users$lat, users$time,
 group_by=users$id, col=c('royalblue', 'orangered'))

invisible(readline(prompt="Press [enter] to continue"))
mobility3d.close()
}

