% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{entropy.space}
\alias{entropy.space}
\title{Mobility entropies}
\usage{
entropy.space(probs)
}
\arguments{
\item{probs}{A vector of probability to find the person in unique locations.}
}
\value{
The entropy value in bits.
}
\description{
In Song's paper, three kinds of entorpies are involved to investigate
the predictability of human mobility, i.e., random entropy $S^{rand}$,
spatial entropy $S^{unc}$ and spatio-temproal entropy $S_i$.
}
\examples{
p <- c(0.1, 0.3, 0.5, 0.1)
entropy.space(p)
}
\references{
Song, C. et al., Limits of predictability in human mobility.
 Science 2010, 1018. (doi:10.1126/science.1177170)
}
\seealso{
\code{\link{entropy.rand}}, \code{\link{entropy.spacetime}}
}
