% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{entropy.spacetime}
\alias{entropy.spacetime}
\title{Mobility entropy}
\usage{
entropy.spacetime(x, y = NULL)
}
\arguments{
\item{x, y}{A chronologically ordered sequence of visited locations. The sequence
is given by a vector (e.g. location names) or two vectors of location coordinates
(e.g., latitude and longitude values).}
}
\value{
The entropy value in bits.
}
\description{
In Song's paper, three kinds of entropy are involved to investigate
the predictability of human mobility, i.e., random entropy \eqn{S^{rand}},
spatial entropy \eqn{S^{unc}} and spatio-temproal entropy \eqn{S_i}.
}
\references{
Song, C. et al., Limits of predictability in human mobility.
 Science 2010, 1018. (doi:10.1126/science.1177170)
}
\seealso{
\code{\link{entropy.rand}}, \code{\link{entropy.space}}
}
