% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euc.dist}
\alias{euc.dist}
\title{Euclidean distance between two points}
\usage{
euc.dist(x1, x2)
}
\arguments{
\item{x1}{First point as a numeric vector}

\item{x2}{Second point as a numeric vector}
}
\value{
The Euclidean distance between the two points
}
\description{
Calculate the Euclidean distance between two points in any dimension.
}
\examples{
euc.dist(c(0,0), c(3,4))
}
