% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdists.R
\name{fit.power.law}
\alias{fit.power.law}
\title{Fit a power-law}
\usage{
fit.power.law(x, y, xmin = min(x), xmax = max(x), plot = TRUE, add = TRUE, ...)
}
\arguments{
\item{x}{A vector of independent variable.}

\item{y}{A vector of dependent variable.}

\item{xmin}{The lower bound point of x.}

\item{xmax}{The higher truncated point of x.}

\item{plot}{Whether to plot the fitted curve.}

\item{add}{Whether to add the fitted curve to current plot.}

\item{...}{Extra parameters to \code{\link{curve}}.}
}
\value{
A list of values for a and lambda.
}
\description{
Model: y ~ a * x^{-lambda}
}
\seealso{
\code{\link[igraph]{fit_power_law}}, \code{\link{fit.truncated.power.law}}
}
