% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{fit_polyexp}
\alias{fit_polyexp}
\title{Fit a poly-exponential distribution}
\usage{
fit_polyexp(x, y, xmin = min(x), xmax = max(x), plot = TRUE, add = TRUE,
  ...)
}
\arguments{
\item{x}{A vector of independent variable.}

\item{y}{A vector of dependent variable.}

\item{xmin}{The lower bound point of x.}

\item{xmax}{The higher truncated point of x.}

\item{plot}{Whether to plot the fitted curve.}

\item{add}{Whether to add the fitted curve to current plot.}

\item{...}{Extra parameters to \code{\link{curve}}.}
}
\value{
A list of values for a, b, c, and d.
}
\description{
Model: y ~ exp(a*x^2 + b*x + c) * x^d
}

