% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{flow.stat}
\alias{flow.stat}
\title{Calculate flow stat between locations}
\usage{
flow.stat(loc, stime, etime, gap = 86400)
}
\arguments{
\item{loc}{A 1D vector to record locations of movement history}

\item{stime}{The starting timestamp (SECONDS) vector of movement history}

\item{etime}{The ending timestamp (SECONDS) vector of movement history}

\item{gap}{The temporal idle interval}
}
\description{
Calculate flow stat between locations
}
\examples{
data(movement)

user_move <- subset(movement, id==1)
sessions <- gen.sessions(user_move[,c("loc", "time")])

res <- with(sessions, flow.stat(loc, stime, etime))
head(res)
}
\seealso{
\code{\link{gen.sessions}}, \code{\link{flowmap}}, \code{\link{flowmap2}},
   \code{\link{plot_flowmap}}
}
