% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowmap.R
\name{flow_stat}
\alias{flow_stat}
\title{Calculate flow stat between locations}
\usage{
flow_stat(loc, stime, etime, gap = 86400)
}
\arguments{
\item{loc}{A 1D vector to record locations of movement history}

\item{stime}{The starting timestamp (SECONDS) vector of movement history}

\item{etime}{The ending timestamp (SECONDS) vector of movement history}

\item{gap}{The temporal idle interval}
}
\description{
Calculate flow stat between locations
}
\examples{
data(movement)

user_move <- subset(movement, id==1)
sessions <- compress_mov(user_move[,c("loc", "time")])

with(sessions, flow_stat(loc, stime, etime))
}
\seealso{
\code{\link{compress_mov}}, \code{\link{flowmap}}, \code{\link{flowmap2}},
   \code{\link{draw_flowmap}}
}

