% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowmap.R
\name{flowmap_stat}
\alias{flowmap_stat}
\title{Calculate flow stat between locations}
\usage{
flowmap_stat(loc, stime, etime, gap = 8 * 3600)
}
\arguments{
\item{loc}{A 1D vector to record locations of movement history}

\item{stime}{The starting timestamp (SECONDS) vector of movement history}

\item{etime}{The ending timestamp (SECONDS) vector of movement history}

\item{gap}{The temporal idle interval}
}
\description{
Calculate flow stat between locations
}

