% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Great Circle Distance (GCD)}
\usage{
gcd(p1, p2, type = "slc")
}
\arguments{
\item{p1}{Location of point 1 with (longitude, latitude) coordinates.}

\item{p2}{Location of point 2 with (longitude, latitude) coordinates.}

\item{type}{Specific algorithm to use, c('slc', 'hf', 'vif').}
}
\value{
Distance in kilometers (km).
}
\description{
Calculates the geodesic distance between two points specified by radian
latitude/longitude using one of the Spherical Law of Cosines (slc),
the Haversine formula (hf), or the Vincenty inverse formula for
ellipsoids (vif).
}
\examples{
# Point in (long, lat) format
p1 <- c(120.0, 30.0)
p2 <- c(120.5, 30.5)

gcd(p1, p2)
gcd(p1, p2, type="hf")
gcd(p1, p2, type="vif")
}

