% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{gen.sessions}
\alias{gen.sessions}
\title{Merge individual ST points into sessions.}
\usage{
gen.sessions(x, y = NULL, t = NULL, gap = 0.5 * 3600)
}
\arguments{
\item{x, y, t}{see params of \code{\link{stcoords}}}

\item{gap}{the time tolerance (sec) to conbime two continuous observations}
}
\description{
Remove duplicate location records in user's movement history.
Continuous records at the same location is merged into a single session
(with interval less than `gap`) recording the starting and ending times.
}
\examples{
data(movement)
library(dplyr)

cmov <- movement \%>\% dplyr::filter(id<10) \%>\%
 group_by(id) \%>\% do(gen.sessions(x=.$loc, t=.$time))
head(cmov)
}
\seealso{
\code{\link{flowmap}}, \code{\link{flowmap2}}, \code{\link{flow.stat}},
   \code{\link{plot.flowmap}}, \code{\link{stcoords}}
}

