% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcoords.R
\name{geo2cart.radian}
\alias{geo2cart.radian}
\title{Convert geopoints in radians to Cartesian coordinates.}
\usage{
geo2cart.radian(x)
}
\arguments{
\item{x}{A 2D vector (lat, long) representing the geo-point in radians.}
}
\description{
Convert geopoints in radians to Cartesian coordinates.
}
\seealso{
\code{\link{cart2geo.radian}}
}

