% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{geo2cart}
\alias{geo2cart}
\title{Geopoint and Cartesian conversion}
\usage{
geo2cart(x)
}
\arguments{
\item{x}{A 2D vector (lat, long) representing the geo-point in degrees.}
}
\value{
A unit-length 3D vector (x, y, z) in Cartesian system.
}
\description{
Converting geo-points in lat/long into Cartesian coordinates.
}
\examples{
geo2cart(c(30, 120))
}
\seealso{
\code{\link{geo2cart.radian}}, \code{\link{cart2geo}}
}
