% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.levels}
\alias{heatmap.levels}
\title{Generate color breaks for heatmap}
\usage{
heatmap.levels(z, nlevels = 10)
}
\arguments{
\item{z}{A numeric vector of values to classify}

\item{nlevels}{The number of levels/breaks to generate}
}
\value{
A numeric vector of break points
}
\description{
Generate color breaks for heatmap visualization using equal interval classification.
}
\examples{
heatmap.levels(rnorm(100), nlevels=10)
}
