% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hour2date}
\alias{hour2date}
\title{Converting UNIX hour to Data object}
\usage{
hour2date(hour, tz = "Asia/Shanghai")
}
\arguments{
\item{hour}{Hours from UNIX epoch}

\item{tz}{The time zone string}
}
\description{
Convert UNIX hour (calculated by dividing UNIX seconds with 3600)
to date at local time zone.
}

