% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hour2tod}
\alias{hour2tod}
\title{Converting UNIX hour to Time-of-Day}
\usage{
hour2tod(hour, tz = "Asia/Shanghai")
}
\arguments{
\item{hour}{Hours from UNIX epoch}

\item{tz}{The time zone string}
}
\description{
Convert UNIX hour (calculated by dividing UNIX seconds with 3600)
to to Time-of-Day (TOD) at local time zone.
}

