% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hour2tow}
\alias{hour2tow}
\title{Converting UNIX hour to Time-of-Week}
\usage{
hour2tow(hour, tz = "Asia/Shanghai")
}
\arguments{
\item{hour}{Hours from UNIX epoch}

\item{tz}{The time zone string}
}
\description{
Convert UNIX hour (calculated by dividing UNIX seconds with 3600)
to to Time-of-Week (TOW) at local time zone.
}

