% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stutils.R
\name{in_area}
\alias{in_area}
\title{Geographic area checking}
\usage{
in_area(lon, lat, area)
}
\arguments{
\item{lon,lat}{The point to be checked.}

\item{area}{The area defined by two points c(lon1, lat1, lon2, lat2).}
}
\description{
Check if the given lon-lat pair falls into specific area.
The area is a 4-length vector with lon-lat pairs of two points that
confine the area boundaries.
}
\examples{
in_area(120.1, 30.1, c(120.0,30.0,120.5,30.5))
}

