% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lonlat2xy}
\alias{lonlat2xy}
\title{Convert longitude/latitude to x/y coordinates}
\usage{
lonlat2xy(lon, lat)
}
\arguments{
\item{lon}{Vector of longitude coordinates}

\item{lat}{Vector of latitude coordinates}
}
\value{
A data frame with x and y columns representing local coordinates
}
\description{
Convert geographic coordinates (longitude/latitude) to local x/y coordinates
using great circle distances from the minimum longitude and latitude points.
}
\examples{
lonlat2xy(c(120, 120.1, 120.2), c(30, 30.1, 30.2))
}
