% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map.R
\name{map3d}
\alias{map3d}
\title{Add a 3D map surface}
\usage{
map3d(upperLeft, lowerRight, h = 0, zoom = NULL,
type = c("osm", "osm-bw", "maptoolkit-topo", "waze", "mapquest", "mapquest-aerial", "bing", "stamen-toner", "stamen-terrain", "stamen-watercolor", "osm-german", "osm-wanderreitkarte", "mapbox", "esri", "esri-topo", "nps", "apple-iphoto", "skobbler", "cloudmade-<id>", "hillshade", "opencyclemap", "osm-transport", "osm-public-transport", "osm-bbike", "osm-bbike-german"),
minNumTiles = 9L, mergeTiles = TRUE)
}
\arguments{
\item{upperLeft}{the upper left lat and long}

\item{lowerRight}{the lower right lat and long}

\item{h}{the horizontal plane to locate the map surface}

\item{zoom}{the zoom level. If null, it is determined automatically}

\item{type}{the tile server from which to get the map}

\item{minNumTiles}{If zoom is null, zoom will be chosen such that the number
of map tiles is greater than or equal to this number.}

\item{mergeTiles}{should map tiles be merged into one tile}
}
\description{
This method add a 3D map surface to the RGL plot. The backend map service is
supported by OpenStreetMap package. All parameters except for h are
consistent with the 'openmap' function in OSM.
}
\examples{
data(movement)
u1 <- subset(movement, id==3)
u1$time <- (u1$time - min(u1$time)) / 3600
region.lat1 <- min(u1$lat) - 0.005
region.lat2 <- max(u1$lat) + 0.005
region.lon1 <- min(u1$lon) - 0.005
region.lon2 <- max(u1$lon) + 0.005

rgl.clear()
rgl.clear("lights")
rgl.bg(color="lightgray")
rgl.viewpoint(theta=30, phi=45)
rgl.light(theta = 45, phi = 45, viewpoint.rel=TRUE)

map3d(c(region.lat2, region.lon1), c(region.lat1, region.lon2),
      min(u1$time), 10, "esri")

axes3d(edges = "bbox", labels = TRUE, tick = TRUE, nticks = 5, box=FALSE,
       expand = 1.03, col="black", lwd=0.8)
}

