% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform-v.R
\name{match_as}
\alias{match_as}
\title{Approximately match}
\usage{
match_as(x, y)
}
\arguments{
\item{x}{A given vector to be matched}

\item{y}{A target vector to calculate absolute approximation}
}
\description{
Match x to y approximately, and return the index of y,
which is mostly near to each value in x.
A variate of match() or %in%
}
\examples{
a <- c(1,2,3)
b <- c(0.1, 0.2, 0.5)
match_as(a, b)
}

