% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint.R
\name{midpoint}
\alias{midpoint}
\title{Geographic midpoint calculation}
\usage{
midpoint(lat, lon, w = rep(1, length(lat)))
}
\arguments{
\item{lat, lon}{The location points}

\item{w}{The weighted value for each point}
}
\value{
The geographic midpoint in lat/lon
}
\description{
Calculate the midpoint given a list of locations denoted by
latitude and longitude coordinates.
}
\examples{
lat <- c(30.2, 30, 30.5)
lon <- c(120, 120.4, 120.5)

# equal weight
midpoint(lat, lon)

# custom weight
w <- c(1, 2, 1)
midpoint(lat, lon, w)

}
\references{
\url{http://www.geomidpoint.com/calculation.html}
}
\seealso{
\code{\link{radius_of_gyration}}
}

