% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minor.ticks.axis}
\alias{minor.ticks.axis}
\title{Adding minor ticks to R basic plots}
\usage{
minor.ticks.axis(ax, n, lab = TRUE, tick.ratio = 0.5, mn, mx, ...)
}
\arguments{
\item{ax}{Which axis to add minor ticks}

\item{n}{The number of minor ticks in each segment}

\item{lab}{If show minor ticks' labels}

\item{tick.ratio}{Ratio of minor ticks' marks}

\item{mn, mx}{The min and max value of minor ticks}

\item{...}{Other parameters for axis() function}
}
\description{
Add minor ticks for log-log plots based on R basic
graphic engine.
}
\concept{graphic extensions}
