% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movr.r
\docType{package}
\name{movr}
\alias{movr}
\title{movr: inspecting human mobility with R}
\description{
A package targeting at analyzing, modeling, and visualizing
human mobility from temporal and spatial perspectives.
}
