% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcorr.R
\name{pairwise.dist}
\alias{pairwise.dist}
\title{Calculate pairwise distances in a 2D region}
\usage{
pairwise.dist(x, y, z)
}
\arguments{
\item{x, y}{the coordinates of given random field}

\item{z}{the observation value for each point (x,y)}
}
\value{
the data frame of c('r', 'z1', 'z2')
}
\description{
Given a random field defined by (x, y) and observation value z, this function
calculates the distance between each location pair and appends original observation
values into the same row. The output can be exploited to calculate spatial correlations.
Note that only Euclidean distance is supported merely for this time.
}
\seealso{
\code{\link{spatial.corr}}
}
