% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{people.occurrence}
\alias{people.occurrence}
\title{Calculate the ocurrence of unique people at each stay point.}
\usage{
people.occurrence(uid, x, y)
}
\description{
Calculate the ocurrence of unique people at each stay point.
}
\examples{
people.occurrence(movement$id, movement$lon, movement$lat)
}

