% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Heatmap of xyz data}
\usage{
\method{plot}{heatmap}(
  x,
  y,
  z,
  nx = 100,
  ny = 100,
  na = 0,
  nlevels = 10,
  levels = NULL,
  colors = rev(brewer.pal(11, "Spectral")),
  bias = 1,
  ...
)
}
\arguments{
\item{x, y}{the coordinates of each z value.}

\item{z}{the observed value for each (x,y) point.}

\item{nx, ny}{the number of bins in x and y dimension, used by \code{\link{vbin.grid}}.}

\item{na}{replacement for NA value in matrix bins, used by \code{\link{vbin.grid}}.}

\item{nlevels}{the number of colorful stages to plot, see `filled.contour`.}

\item{levels}{a numeric vector to indicate the colorful stages, see `filled.contour`.}

\item{colors}{a vector of color names or hex values, see `colorRampPalette`.}

\item{bias}{a positive number. Higher values give more widely spaced colors at the high end, see `colorRampPalette`.}

\item{...}{other parameters sent to `filled.contour`.}
}
\description{
Visualize the xyz data in a 2d heatmap.
}
