% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{plot.traj.graph}
\alias{plot.traj.graph}
\title{Visualize individual's trajectories.}
\usage{
plot.traj.graph(loc, time, ...)
}
\description{
This funciton plots a trajectory of a specific user as a weighted
graph. Each node represents a stay point whose size indicates the
(log) lenght of dwelling time. Each directed edge means the existence
transition between stay points and its width indicates the (log)
transition frenquencies.
}

