% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{plot.traj3d}
\alias{plot.traj3d}
\title{Visualize trajectories using RGL 3D.}
\usage{
plot.traj3d(x, y, t, group_by = NULL, col = NULL, xlab = "", ylab = "",
  tlab = "", ...)
}
\arguments{
\item{x, y}{Numeric vectors of spatial coordinates}

\item{t}{The temporal vector for each (x,y) point.}

\item{group_by}{A group indicator when multiple users are visualized.}

\item{col}{A vector of color strings. It must have the same length as unique(group_by).}

\item{xlab, ylab, tlab}{The labels for each axis.}

\item{...}{Other parameters for \code{\link[rgl]{plot3d}} or \code{\link[rgl]{axes3d}}}
}
\description{
Visualize trajectories using RGL 3D.
}
\examples{
data(movement)

users <- subset(movement, id \%in\% c(23, 20)) \%>\%
 mutate(time = time/86400 - min(time/86400)) \%>\%
 dplyr::filter(time <= 30)
\dontrun{
plot.traj3d(users$lon, users$lat, users$time,
 group_by=users$id, col=c('royalblue', 'orangered'))

invisible(readline(prompt="Press [enter] to continue"))
traj3d.close()
}
}

