% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{plot_traj3d}
\alias{plot_traj3d}
\title{Visualize trajectories using RGL 3D.}
\usage{
plot_traj3d(
  x,
  y,
  t,
  group_by = NULL,
  col = NULL,
  xlab = "",
  ylab = "",
  tlab = "",
  ...
)
}
\arguments{
\item{x, y}{Numeric vectors of spatial coordinates}

\item{t}{The temporal vector for each (x,y) point.}

\item{group_by}{A group indicator when multiple users are visualized.}

\item{col}{A vector of color strings. It must have the same length as unique(group_by).}

\item{xlab, ylab, tlab}{The labels for each axis.}

\item{...}{Other parameters for \code{\link[rgl]{plot3d}} or \code{\link[rgl]{axes3d}}}
}
\description{
Visualize trajectories using RGL 3D.
}
\note{
This function requires the \pkg{rgl} package to be installed.
You can install it with \code{install.packages("rgl")}.
}
\examples{
data(movement)

users <- subset(movement, id \%in\% c(23, 20))
users <- dplyr::mutate(users, time = time/86400 - min(time/86400))
users <- dplyr::filter(users, time <= 30)
\dontrun{
plot_traj3d(users$lon, users$lat, users$time,
 group_by=users$id, col=c('royalblue', 'orangered'))

invisible(readline(prompt="Press [enter] to continue"))
traj3d.close()
}
}
