% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{plot_traj_graph}
\alias{plot_traj_graph}
\title{Visualize individual's trajectories.}
\usage{
plot_traj_graph(loc, time, ...)
}
\description{
This function plots a trajectory of a specific user as a weighted
graph. Each node represents a stay point whose size indicates the
(log) length of dwelling time. Each directed edge means the existence
transition between stay points and its width indicates the (log)
transition frequencies.
}
