% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquality.R
\name{Q.iovan}
\alias{Q.iovan}
\title{Spatiotemporal data quality indicators of Iovan}
\usage{
Q.iovan(x, y, t)
}
\arguments{
\item{x, y, t}{input for \code{\link{stcoords}}. If x and y represent longitude and latitude respectively,
please make sure that longitude is located as the first param.}
}
\value{
theta the speed index

U the uncertainty

Q the unified quality indicator for each data point

H the entropic quality indicator for each user
}
\description{
In the paper, Iovan et al.[1] addressed user sampling issues from local and global
aspects, respectively. The local measures quantify the quality of each observation
point for each individual, which include speed index (theta), uncertainty (U),
and a unified quality indicator (Q) based of previous two metrics. The global measure (H)
quantifies the data quality for a whole trajectory, by considering the information entropy
of a vector of quality indicators (Q).
}
\examples{
user <- movement \%>\% dplyr::filter(id==1)
Q.iovan(user$lon, user$lat, user$time/3600)

}
\references{
[1] C. Iovan, A.-M. Olteanu-Raimond, T. Couronne, and Z. Smoreda,
"Moving and Calling: Mobile Phone Data Quality Measurements and Spatiotemporal Uncertainty in Human Mobility Studies," in
Geographic Information Science at the Heart of Europe, 2013, pp. 247–265.
}

