% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcoords.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees.}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{A number or vector of radians}
}
\description{
Convert radians to degrees.
}
\seealso{
\code{\link{deg2rad}}
}

