% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiusgyration.R
\name{radius_gyration}
\alias{radius_gyration}
\title{Radius of gyration for human mobility}
\usage{
radius_gyration(lat, lon, w = rep(1, length(lat)))
}
\arguments{
\item{lat,lon}{The geographic coordinates of locations}

\item{w}{The weight value for each location}
}
\value{
The radius of gyration (km)
}
\description{
Given a series of locations denoted by lat/lon coordinates,
the radius of gyration for individual is calculated.
}
\examples{
lat <- c(30.2, 30, 30.5)
lon <- c(120, 120.4, 120.5)
radius_gyration(lat, lon)
}
\references{
M. C. Gonzalez, C. A. Hidalgo, and A.-L. Barabasi,
"Understanding individual human mobility patterns,"
Nature, vol. 453, no. 7196, pp. 779-782, Jun. 2008.
}
\seealso{
\code{\link{midpoint}}
}

