% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{Rcolors}
\alias{Rcolors}
\title{R Colors}
\usage{
Rcolors()
}
\description{
Plot matrix of R colors, in index order, 25 per row.
This is for convenient reference when programming.
}
\references{
\link{http://research.stowers-institute.org/efg/R/Color/Chart/}
}

