% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{rep_each}
\alias{rep_each}
\title{Replicate elements of vector}
\usage{
rep_each(x, times = 2)
}
\arguments{
\item{x}{a vector}

\item{times}{the number of replication times of each element.}
}
\description{
This is a slight modification of \code{rep} in basic package. It replicates
each element of a vector one by one to construct a new vector.
}
\examples{
rep(1:10, 2)
rep_each(1:10, 2)
}
\seealso{
\code{\link{seq_approximate}},
}

