% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared Error (RMSE)}
\usage{
RMSE(x, y)
}
\arguments{
\item{x}{A given vector to calculate RMSE.}

\item{y}{The target vector}
}
\description{
calculate the root mean squared error (RMSE) of two vectors.
}
\examples{
RMSE(c(1,2,3,4), c(2,3,2,3))
}

