% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rot90}
\alias{rot90}
\title{Rotate a matrix by 90 degrees}
\usage{
rot90(m)
}
\arguments{
\item{m}{A matrix to rotate}
}
\value{
The rotated matrix
}
\description{
Rotate a matrix by 90 degrees clockwise.
Rotate matrix 90 degrees clockwise
}
\details{
Rotate a matrix 90 degrees clockwise by transposing and reversing columns.
}
\examples{
m <- matrix(1:9, nrow=3)
rot90(m)
}
