% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_approximate}
\alias{seq_approximate}
\title{Approximately matching sequence}
\usage{
seq_approximate(x, y)
}
\arguments{
\item{x}{A given vector to be matched}

\item{y}{A target vector to calculate absolute approximation}
}
\description{
Match x to y approximately, and return the index of y,
which is mostly near to each value in x.
A variate of match() or %in%
}
\examples{
a <- c(1,2,3)
b <- c(0.1, 0.2, 0.5)
seq_approximate(a, b)
}
\seealso{
\code{\link{seq_along}}, \code{\link{rep_each}}
}

