% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_collapsed}
\alias{seq_collapsed}
\title{Sequencing by collapsing adjacent same values}
\usage{
seq_collapsed(v)
}
\arguments{
\item{v}{The input vector.}
}
\description{
Generate integer sequence by assigning the same adjacent values to the same
level.
}
\examples{
seq_collapsed(c(1,2,2,3,2,2))
}
\seealso{
\code{\link{seq_along}}, \code{\link{seq_distinct}},
   \code{\link{vbin}}, \code{\link{vbin_range}}, \code{\link{vbin_grid}}
}

