% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequences.R
\name{standardize}
\alias{standardize}
\title{Vector Normalization}
\usage{
standardize(x)
}
\arguments{
\item{x}{A vector to be normalized.}
}
\description{
Normalize a given vector.
}
\examples{
standardize(c(1,2,3,4,5,6))
}
\seealso{
\code{\link{standardize_st}}
}

