% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{standardize_st}
\alias{standardize_st}
\title{Normalization over spatial and temporal scale}
\usage{
standardize_st(scoord, tcoord, value, alpha = 0.5)
}
\arguments{
\item{scoord}{a 1D vector of spatial coordinate}

\item{tcoord}{a 1D vector of temporal coordinate}

\item{value}{a value vector for each (scoord, tcoord)}

\item{alpha}{a tuning parameter controling the weight of space and time}
}
\description{
Scale the value along spatial and temporal coordinates simultaneously.
}
\examples{
scoord <- rep(seq(6), 2)
tcoord <- rep(c(1,2), each=6)
value <- runif(6 * 2)
standardize_st(scoord, tcoord, log10(1+value))
}

