% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{stcoords}
\alias{stcoords}
\title{Spatiotemporal data formatting}
\usage{
stcoords(x, y = NULL, t = NULL)
}
\arguments{
\item{x}{A vector, data frame or matrix.}

\item{y}{A vector.}

\item{t}{A vector.}
}
\description{
Format spatiotemporal series in a unified manner for both 1D and 2D locations.
If \code{x} is a data frame or matrix, \code{y} and \code{t} are omitted.
}
\details{
If \code{x} is a data frame (3 columns), this function automatically identify spatial
and temporal values by column names, i.e., (x,y,t) and (lat,lon,time).
Otherwise, the column indexes are employed as [, 1] and [, 2] being the space
coordinates and [, 3] being the timestamps.

If \code{x} is a data frame (2 columns), similar policies are involved, but
alternatively column names (x, t) and (loc, time) are used.

If \code{x} is a matrix, column indexes are used merely.

If \code{x} is a vector, dimensions of space coordinates are determined
by both \code{x} and \code{y}, and the time dimension by \code{t}.
}
\examples{
## One data frame with columes x, y, t
x <- data.frame(x=rep(1:10, 2), y=rep_each(1:10, 2), t=1:20)
stcoords(x)

## One data frame without demanded colume names
x <- data.frame(rep(1:10, 2), rep_each(1:10, 2), 1:20)

## One data frame with two columes loc, time
x <- data.frame(loc=rep(1:10, 2), time=1:20)

## With vectors
stcoords(x=rep(1:10, 2), t=1:20)
}
\seealso{
\code{\link{stcoords_1d}}
}

