% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{stcoords_1d}
\alias{stcoords_1d}
\title{Spatiotemporal data formatting (1D)}
\usage{
stcoords_1d(x, y = NULL, t = NULL)
}
\arguments{
\item{x, y, t}{params of \code{stcoords}}
}
\description{
Similar to \code{\link{stcoords}}, return location instead of x, y coordinates.
}
\examples{
x <- data.frame(rep(1:10, 2), rep_each(1:10, 2), 1:20)
stcoords_1d(x)
}
\seealso{
\code{\link{stcoords}}
}

