% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{traj3d.close}
\alias{traj3d.close}
\title{Close RGL 3D device}
\usage{
traj3d.close()
}
\description{
Close the current RGL 3D device. This is a convenience function
that checks if the rgl package is available before closing.
}
\note{
This function requires the \pkg{rgl} package to be installed.
You can install it with \code{install.packages("rgl")}.
}
\examples{
\dontrun{
# After creating a 3D plot with plot_traj3d
traj3d.close()
}
}
