% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_of_gyration.R
\name{turbo.radius.of.gyration}
\alias{turbo.radius.of.gyration}
\title{Turbo Radius of gyration for human mobility}
\usage{
turbo.radius.of.gyration(lat, lon, w = rep(1, length(lat)))
}
\description{
Turbo Radius of gyration for human mobility
}
