% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{vbin.grid}
\alias{vbin.grid}
\title{2D random field binning}
\usage{
vbin.grid(x, y, z, nx, ny, FUN = mean, na = NA)
}
\arguments{
\item{x, y, z}{a random field with location vectors (x, y) and value vector z.
They must have the same length.}

\item{nx, ny}{the number of bins in x and y dimension.}

\item{FUN}{a function to calculate statistics in each 2D bin.}

\item{na}{Replacement for NA value in matrix bins.}
}
\value{
a matrix with row (column) names being the center points of x (y) dim,
and with cell value being the aggregate statistics calculated by FUN.
}
\description{
Generate a bined matrix given a 2D random field.
}
\examples{
vbin.grid(1:20, 20:1, runif(20), nx=5, ny=5)
}
\seealso{
\code{\link{seq_approximate}}, \code{\link{vbin}}, \code{\link{vbin.range}}
}
