% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{vbin.range}
\alias{vbin.range}
\title{Vector range binning}
\usage{
vbin.range(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the number of bins}
}
\value{
the center of each interval
}
\description{
Bin the range of given vector into n itervals.
}
\examples{
vbin.range(10:20, 3)
}
\seealso{
\code{\link{seq_approximate}}, \code{\link{vbin}}, \code{\link{vbin.grid}}
}
