% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequences.R
\name{vbin_grid}
\alias{vbin_grid}
\title{2D random field binning}
\usage{
vbin_grid(x, y, z, nx, ny, FUN = mean, na = NA)
}
\arguments{
\item{x,y,z}{a random field with location vectors (x, y) and value vector z.
They must have the same length.}

\item{nx,ny}{the number of bins in x and y dimension.}

\item{FUN}{a function to calculate statistics in each 2D bin.}

\item{na}{Replacemnet for NA value in matrix bins.}
}
\value{
a matrix with row (column) names being the center points of x (y) dim,
and with cell value being the aggregate statistics calculated by FUN.
}
\description{
Generate a bined matrix given a 2D random field.
}
\examples{
vbin_grid(1:20, 20:1, runif(20), nx=5, ny=5)
}
\seealso{
\code{\link{match_as}}, \code{\link{vbin}}, \code{\link{vbin_range}}
}

