% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequences.R
\name{vbin_range}
\alias{vbin_range}
\title{Vector range binning}
\usage{
vbin_range(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the number of bins}
}
\value{
the center of each interval
}
\description{
Bin the range of given vector into n itervals.
}
\examples{
vbin_range(10:20, 3)
}
\seealso{
\code{\link{vbin}}
}

