% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{voronoi2polygons}
\alias{voronoi2polygons}
\title{Voronoi to polygon}
\usage{
voronoi2polygons(x, poly)
}
\description{
Convert Voronoi diagram generated by `deldir` package into SpatialPolygons
http://stackoverflow.com/questions/12156475/combine-voronoi-polygons-and-maps/12159863#12159863
}

