% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_layers.R
\name{voronoi3d}
\alias{voronoi3d}
\title{3D voronoi canvas for RGL}
\usage{
voronoi3d(
  x,
  y,
  group_by = NULL,
  col = NULL,
  side = "y",
  col.seg = "grey",
  lty = 1,
  lwd = 1
)
}
\description{
3D voronoi canvas for RGL
}
