% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mobility.R
\name{cal_place_dwelling}
\alias{cal_place_dwelling}
\title{Calculate dwelling period by averaging dwelling
time between different consecutive locations}
\usage{
cal_place_dwelling(loc, time)
}
\arguments{
\item{loc}{A vector of location identifiers}

\item{time}{A vector of timestamps corresponding to each location}
}
\value{
A data frame with columns 'loc' and 'dwelling' containing the total dwelling time for each location
}
\description{
Calculate dwelling period by averaging dwelling
time between different consecutive locations
}
