% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point}
\alias{dq.point}
\title{Spatiotemporal data quality of stay points}
\usage{
dq.point(stcoords, pc, po, dq.min = 1e-05, na = dq.min)
}
\arguments{
\item{stcoords}{the spatiotemporal coordinates, see \code{\link{stcoords}}}

\item{pc}{the point coverage, see \code{\link{point.coverage}}}

\item{po}{the occurrence of people, see \code{\link{people.occurrence}}}

\item{dq.min}{the lower bound of data quality. Values less than this bound
will be set forcely to this quality (default 1e-5)}

\item{na}{the replacement of NA (default dq.min)}
}
\value{
the origianl sessions with extra fields: c('dq.s', 'dq.d', 'dq'). `dq.s`
 and `dq.d` give the STATIC and DYNAMIC quality measures respectively, while
 `dq` is the Harmonic mean of previous two.
}
\description{
This funciton gives the local quality metric of spatiotemporal data points.
The metric contains two parts of quality information, ie STATIC and DYNAMIC.
The STATIC part comes from the spatial and temporal coverage of each point,
which quantifies the uncertainty of user whereabout caused by the intrinsic
limitation of measurement method (eg location coverage). The DYNAMIC part
gives the quanlity from dynamic information like speed, which emploies the
Iovan's algorithm as addressed in \code{\link{dq.iovan}}.
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==1)
pc <- point.coverage(movement$lon, movement$lat)
po <- people.occurrence(movement$id, movement$lon, movement$lat)

stc <- stcoords(u1[,c('lon','lat','time')])
head(dq.point(stc, pc, po))

sessions <- gen.sessions(stc$x, stc$y, stc$t)
head(dq.point2(sessions, pc, po))
}
\seealso{
\code{\link{dq.point2}}
}
