% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point}
\alias{dq.point}
\title{Spatiotemporal data quality of stay points}
\usage{
dq.point(x, y, t, pc, po)
}
\description{
Spatiotemporal data quality of stay points
}
\examples{
u1 <- movement \%>\% dplyr::filter(id==1)
pc <- point.coverage(u1$lon, u1$lat)
po <- people.occurrence(movement$id, movement$lon, movement$lat)
head(dq.point(u1$lon, u1$lat, u1$time, pc, po))
}

