% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.point2}
\alias{dq.point2}
\title{Spatiotemporal data quality of stay points}
\usage{
dq.point2(sessions, pc, po, dq.min = 1e-05, na = dq.min)
}
\arguments{
\item{sessions}{a data.frame of session data generated by \code{\link{gen.sesions}}
or external tools. Four columns are required: c('stime', 'etime', 'x', 'y')}

\item{pc}{the point coverage, see \code{\link{point.coverage}}}

\item{po}{the occurrence of people, see \code{\link{people.occurrence}}}

\item{dq.min}{the lower bound of data quality. Values less than this bound
will be set forcely to this quality (default 1e-5)}

\item{na}{the replacement of NA (default dq.min)}
}
\value{
as the return of \code{\link{dq.point}}
}
\description{
This is the inner implementaion of \code{\link{dq.point}}. Differently,
this function accepts the session data generated by \code{\link{gen.sessions}},
or external utilities.
}
\seealso{
\code{\link{dq.point}}
}

