% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{dq.traj2}
\alias{dq.traj2}
\title{Spatiotemporal data quality of user trajectories}
\usage{
dq.traj2(dqPoints)
}
\description{
Spatiotemporal data quality of user trajectories
}
\seealso{
\code{\link{dq.traj}}
}

