% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{draw_flowmap}
\alias{draw_flowmap}
\title{Visualize flowmap.}
\usage{
draw_flowmap(from_lat, from_lon, to_lat, to_lon, dist.log = TRUE,
  weight = NULL, weight.log = TRUE, gc.breaks = 5, col.pal = c("white",
  "blue", "black"), col.pal.bias = 0.3, col.pal.grad = 200,
  new.device = TRUE, bg = "black", ...)
}
\arguments{
\item{from_lat}{The latitude coordinates of departing point for mobile transitions.}

\item{from_lon}{The longitude coordinates of departing point for mobile transitions.}

\item{to_lat}{The latitude coordinates of arriving point for mobile transitions.}

\item{to_lon}{The longitude coordinates of arriving point for mobile transitions.}

\item{dist.log}{Whether using log-scale distance for line color.}

\item{weight}{The user-defined weight for line color. Larger weight corresponds to lefter color of col.pal.}

\item{weight.log}{Whether using log-scale weight for line color.}

\item{gc.breaks}{The number of intermediate points (excluding two ends) to draw a great circle path.}

\item{col.pal}{A color vector used by \code{colorRampPalette}; must be a valid argument to \code{col2rgb}.
Refer to \url{colorbrewer2.org} to derive more palettes.}

\item{col.pal.bias}{The bias coefficient used by \code{colorRampPalette}. Higher values give more widely
spaced colors at the high end.}

\item{col.pal.grad}{The number of color grades to diffeciate distance.}

\item{new.device}{Whether creating a new device for current plot. Set this parameter as FALSE when
trying to plot multiple flowmaps in one figure.}

\item{bg}{The background color for current plot. It is working when new.device is TRUE.}

\item{...}{Extra parameters for basic plot() function.}
}
\description{
Visualize the mobility statistics (flowmap) from data. Each row in the data
will generate a line on the map.
}
\seealso{
\code{\link{compress_mov}}, \code{\link{flowmap}}, \code{\link{flowmap2}},
   \code{\link{flow_stat}}
}

