% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{entropy.rand}
\alias{entropy.rand}
\title{Major concepts, tools and visualization in the predictability paper.
 Song, C. et al., Limits of predictability in human mobility.
 Science 2010, 1018. (doi:10.1126/science.1177170)
Mobility entropy}
\usage{
entropy.rand(N)
}
\arguments{
\item{N}{The number of unique locations.}
}
\value{
The entropy value in bits.
}
\description{
In Song's paper, three kinds of entropy are involved to investigate
the predictability of human mobility, i.e., random entropy \eqn{S^{rand}},
spatial entropy \eqn{S^{unc}} and spatio-temporal entropy \eqn{S_i}.
}
\references{
Song, C. et al., Limits of predictability in human mobility.
 Science 2010, 1018. (doi:10.1126/science.1177170)
}
\seealso{
\code{\link{entropy.space}}, \code{\link{entropy.spacetime}}
}
