% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowmap.R
\name{flowmap2}
\alias{flowmap2}
\title{Generate flowmap from movement data}
\usage{
flowmap2(uid, loc, stime, etime, gap = 86400)
}
\arguments{
\item{uid}{a vector to record user identities}

\item{loc}{a 1D vector to record locations of movement history}

\item{stime,etime}{compressed session time at each location}

\item{gap}{the maximum dwelling time to consider a valid move between locations}
}
\value{
a data frame with four columns: from, to, total, unique (users)
}
\description{
Use historical movement data to generate flowmap, which records mobility
statistics between two locations 'from' and 'to'.
}
\details{
Different from \code{flowmap}, compressed movement history is used to
generate flow statistics.
}
\seealso{
\code{\link{compress_mov}}, \code{\link{flowmap}}, \code{\link{flow_stat}},
   \code{\link{draw_flowmap}}
}

