% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{gcd}
\alias{gcd}
\title{Great Circle Distance (GCD)}
\usage{
gcd(p1, p2, type = "hf", na = 0)
}
\arguments{
\item{p1}{Location of point 1 with (lat, long) coordinates.}

\item{p2}{Location of point 2 with (lat, long) coordinates.}

\item{type}{Specific algorithm to use, c('slc', 'hf', 'vif').}

\item{na}{Value to return for NA distances}
}
\value{
Distance in kilometers (km).
}
\description{
Calculates the geodesic distance between two points specified by radian
latitude/longitude using one of the Spherical Law of Cosines (slc),
the Haversine formula (hf), or the Vincenty inverse formula for
ellipsoids (vif).
}
\examples{
# Point in (lat, long) format
p1 <- c(30.0, 120.0)
p2 <- c(30.5, 120.5)

gcd(p1, p2)
gcd(p1, p2, type="hf")
gcd(p1, p2, type="vif")
}
\references{
\url{
https://www.r-bloggers.com/2010/11/great-circle-distance-calculations-in-r/}
}
