% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{gen.sessions}
\alias{gen.sessions}
\title{Detect staying sessions of user}
\usage{
gen.sessions(x, y = NULL, t = NULL, gap = 0.5 * 3600, unite.sep = "_")
}
\arguments{
\item{x, y, t}{see parameters of \code{\link{stcoords}}}

\item{gap}{the time tolerance (sec) to combine two continuous observations}

\item{unite.sep}{a separator to combine x and y coordinators into one column,
see also \code{\link{stcoords}}}
}
\description{
This function removes duplicate location records in user's mobility trajectory.
Successive records at the same location are merged into a single session
(with interval less than `gap`) recording the starting and ending times.
}
\examples{
data(movement)
u1 <- movement \%>\% dplyr::filter(id==2)

## 1-column location indicators
head(gen.sessions(u1$loc, t=u1$time))

## 2-column location coordinates
head(gen.sessions(u1$lon, u1$lat, u1$time))
}
\seealso{
\code{\link{flowmap}}, \code{\link{flowmap2}}, \code{\link{flow.stat}}
}
