% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transform-coords.R
\name{geo2cart}
\alias{geo2cart}
\title{Geopoint and Cartesian conversion}
\usage{
geo2cart(x)
}
\arguments{
\item{x}{A 2D vector (lat, long) representing the geo-point in degree.}
}
\value{
A unit-length 3D vector (x, y, z) in Cartesian system.
}
\description{
Converting geo-points in lat/long into Cartesian coordinates.
}
\examples{
geo2cart(c(30, 120))
}

