% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{point.coverage}
\alias{point.coverage}
\title{Calculate the coverage of each stay point according to Voronoi tesselation.}
\usage{
point.coverage(x, y)
}
\description{
Calculate the coverage of each stay point according to Voronoi tesselation.
}
\examples{
head(point.coverage(movement$lon, movement$lat))
}

