% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Rcolors}
\alias{Rcolors}
\alias{Rcolours}
\title{R Colors}
\usage{
Rcolors(huesort=TRUE)
}
\arguments{
\item{huesort}{Boolean value to control ordering by HUE.}
}
\description{
Plot matrix of R colors, in index order, 25 per row.
This is for quick reference when programming.
}
\details{
Copyright: Earl F. Glynn
}
\references{
http://research.stowers-institute.org/efg/R/Color/Chart/
}
