% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transform-v.R
\name{seq_collapsed}
\alias{seq_collapsed}
\title{Sequencing by collapsing adjacent same values}
\usage{
seq_collapsed(v)
}
\description{
Generate integer sequence by assigning the same adjacent values to the same
level.
}
\examples{
seq_collapsed(c(1,2,2,3,2,2))
}

