% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transform-v.R
\name{vbin}
\alias{vbin}
\title{Vector binning}
\usage{
vbin(x, n, center=c(TRUE, FALSE))
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the number of bins}

\item{center}{indication of representing intervals as index (default) or
center points.}
}
\value{
Sequence with interval index or center points.
}
\description{
Bin a vector into `n` intervals in regard with its value range.
The vector x is split into n bins within [min(x), max(x)],
and bin index is given by checking the bin [bin_min, bin_max)
into which data points in x fall.
}
\examples{
vbin(1:10, 3)
vbin(1:10, 3, TRUE)
}

