% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maputils.R
\name{voronoi2polygons}
\alias{voronoi2polygons}
\title{http://stackoverflow.com/questions/12156475/combine-voronoi-polygons-and-maps/12159863#12159863}
\usage{
voronoi2polygons(x, poly)
}
\description{
http://stackoverflow.com/questions/12156475/combine-voronoi-polygons-and-maps/12159863#12159863
}

