% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_layers.R
\name{voronoi3d}
\alias{voronoi3d}
\title{3D voronoi canvas for RGL}
\usage{
voronoi3d(
  x,
  y,
  group_by = NULL,
  col = NULL,
  side = "y",
  col.seg = "grey",
  lty = 1,
  lwd = 1
)
}
\arguments{
\item{x, y}{Coordinates of points}

\item{group_by}{Grouping variable for points}

\item{col}{Colors for each group}

\item{side}{Which side to project the voronoi diagram ('x', 'y', or 'z')}

\item{col.seg}{Color for line segments}

\item{lty}{Line type}

\item{lwd}{Line width}
}
\description{
3D voronoi canvas for RGL
}
\note{
This function requires the \pkg{rgl} package to be installed.
You can install it with \code{install.packages("rgl")}.
}
